refactor: Move test
